requirements.txt
torch
torchvision
numpy
pandas
segmentation-models-pytorch
tqdm
Pillow
scikit-image
shapely
joblib
contextlib2
pathlib
psutil
pyparsing
importlib_resources
